requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
python-dotenv==1.0.0
python-dateutil==2.8.2
tqdm==4.65.0
openpyxl==3.1.2